Separate sizes and fonts for names and lengths.
These changes are in preparation for using the anchor point.

We let the user select the details of the text differently for
names and lengths (as we plan to do for other texts/properties).

There is a bit of cosmetics and comments for where to change
the code in the future so we do not make unnecessary calls
to the backend.