Only consider a box small if the y dimension is small.
The x dimension can get very small easily for leaves when we are
zooming around, and then it always looks as if there are further
collapsed nodes, which is not the case for leaves.

This change avoids it and keeps the check for being small
still basically as useful as before.