Use the anchor point to re-scale elements.
This change completes the functionality of anchor points.

We use them for leaf names so they align in the middle, no
matter what text size they have, and we use them for lengths
so they stay above the branch line. We can now use them for
anything else we want to anchor.

Also, we add dynamic vertical padding to the names. With the
anchoring point working, they re-scale correctly.