Simpler drawers. This is a big change -- see notes below.
Drawers now have information about their type (TYPE="rect"
or "circ") as a class property, and about how many panels they
can draw to (NPANELS). Drawing for panel=0 draws the main
content (with the skeleton of the tree and more), and for any
other panel>0 it draws things that appear aligned with the main
content (in new divs in the frontend or bigger circles). The
Drawers can choose how many panels draw to, and with which
content.

The server exposes the new characteristics of the drawers by
having new endpoints, /drawers (gives the list of drawers,
instead of the old /trees/drawers) and /drawers/<name> (gives the
type ("rect" or "circ") and the number of panels it draws
to). This also avoids having to check in the frontend for the
name of the drawer to see if they are circular and/or if they
have aligned content (and also allows to have aligned content to
multiple panels).

Some changes in naming:
  self.boxes -> self.nodeboxes  (they were always for nodes)
  cone -> outlines (cones were really used as outlines)
  Drawer{Rect,Circ}Simple -> Drawer{Rect,Circ}
  ["box", ...] -> ["nodebox", ...]
  update_tree -> draw_tree

Simplify functions for drawing. Instead of draw_content_inline()
and draw_content_float() we now have a single draw_node(), which
behaves a bit like draw_content_float() before, but also gets the
branch dy (bdy) like draw_content_inline() used to do (and like
it also gets only called if the content of the node is visible).

The separation was useful long time ago, but with the current way
of drawing the nodes in the main and aligned panels it made no
sense anymore. With this change all the drawing logic for new
users is quite easy:
  draw_node - called for a node that needs to be drawn
  draw_collapsed - called for a group of collapsed nodes

The in_viewport() checks for aligned content only care about the
y-dimension. It is as if the viewport was stretched infinitely in
the x-dimension. We tried to do it before too, but the check was
clumsy and failed in certain cases. Now we separate the
y-dimension and check on it, which is what we really want (and it
always works).

We remove the intermediate Drawers like DrawerRectLeafNodes and
substitute them for functions (generators, really) that can be
picked up individually by any Drawer.

For example:
  DrawerRectLeafNodes -> draw_rect_leaf_names
and a drawer can use it like: yield from draw_rect_leaf_names(...)

This hopefully will allow a more composable building of the
drawers. Very close to what "faces" are in the original ete.

There are two new drawers: DrawerCircAlignNames and
DrawerCircAlignHeatmap. They exploit the new way of dealing with
panels to draw the aligned region in the same svg as the main
drawing, but starting where the radius (r) of the tree ends.

Also, many changes in the frontend to accomodate for the new
structure.