Show only the visible angular part in aligned panels (when traversing the -x axis).
There was a problem of overaccepting graphical elements in
aligned panels in circular mode.

Our annular sectors only start at an angle -pi (or more), and
there is no continuous representation of an angular segment
that crosses the -x axis (the angle goes from ~ -pi to ~ +pi).
The check to see if circular elements that are going to be shown
in aligned panels, when the viewport traverses the -x axis, was
true for all angles, so we drew many unnecessary things (and
the interface was slow in this quite common case!).

With this fix, we separate the viewport in two rectangles if it
crosses the -x axis, and then check for the circular graphical
elements if they are contained within those (continuous this time!)
angles.

This seems to be the correct solution. It also goes significantly
faster for those cases now.