Smooth move and zoom the aligned content in circular mode.
The aligned content in circular mode is added to the same
svg as the main content, as a new group (svg "g"), which
is added as another children of the svg.

Thus, when we move and zoom smoothly before making the api
call to redraw, we can apply the same transformation for that
group as we do for the main one.

And it looks cool :)