Remove TODOs about avoiding api calls, they are quite unnecessary.
It is true that we could avoid some redrawing that uses relatively
expensive api calls. For changing the padding and max size of fonts
and so on, for different styles (on names, lengths, etc.).

But:
  * The calls are fast enough in practice (typically hundreds of ms).
  * The code that avoids them would be significanlty more complex
    (not terribly so, but way more than calling update()!).
  * Having those parts with commented code reduces readability of
    the file. A quite unnecessary cognitive burden.

If in the future we want to do something like that, the code could
be similar to:

  const padding = view.name.padding.left / 100;
  Array.from(div_tree.getElementsByClassName("name")).forEach(t => {
      const fs = Number(t.getAttribute("font-size").slice(0, -2));
      const x0 = Number(t.getAttribute("data-x0"));
      t.setAttribute("x", x0 + padding * fs);
  });

and

  Array.from(div_tree.getElementsByClassName("name")).forEach(t => {
      const box = t.getAttribute("data-box").split(",").map(Number);
      const anchor = t.getAttribute("data-anchor").split(",").map(Number);
      ...
  });

and so on.

Note that the code above is not correct for all cases, though. It
is just an example.