Add labels. And quite a bit more -- see below.
Labels are implemented as Drawers that in their draw_node() (and
possibly draw_collapsed()) they use the new parameter sent to the
constructor, "labels", a list of functions that, given a node and
a few more things, return a box and an anchor point and a text to
draw (and a name that identifies the label used to create it, so
we can then manipulate it by selecting its class in the frontend
and modifying its style as if with css).

All the drawers except the basic (Rect and Circ) now derive from
RectLabels and CircLabels. Instead of Drawer*Full we now use
Drawer*LeafNames, which use the labels except for the leaf
names (and collapsed names), which we do better for the
moment "by hand" in this drawer.

The labels are passed to the server as a list of lists with
[expression, nodetype, position] , all json-encoded.

--

The nodeboxes (boxes representing the node in the gui) are now in
the background, so when they are not transparent, they do not
hide the rest of the tree. We copy them into foreground boxes,
which are always transparent, and allow to select the nodes and
use their contextmenus, etc.

--

Separate the datgui menu in 3 menus. The tags, searches, and now
labels too, can become very long, and it is tedious to use a long
menu.

--

Add endpoint "/properties", to retrieve the available properties
defined in any node of a tree.

--

Add endpoint "/reload", to reload a tree (in case we had changed
its structure from the gui).

--

Draw negative x-axis when the circular view goes beyond -pi to
pi. This gives feedback to the user, to let her know that there
is part of the tree disappearing there.

--

Fix drawing of outline in circular mode when angles beyond -pi to
pi.

--

Fix elements beyond -pi to pi appearing on the minimap. It was
caused because viewport is None for the minimap (full plane is
visible), but then we did not check for elements beyond valid
angles.

--

Removed unnecessary self.zoom from get_rect (we used it when we
computed the font size in the backend, but now we send a box and
an anchor point, and it is the frontend that computes the actual
font size)

--

Avoid drawing some lengthlines and childrenlines if out of
view (minor).

--

Quite a few cosmetics and other minor changes too.