Add manually collapsed nodes. And change node_id to a tuple.
We can now select from the gui which nodes we want to keep
always collapsed.

They appear on a new menu, so we can restore them.

Also, the node_id is now a tuple instead of a list, which is a
hashable element and makes more sense (and we can use in checks
like "if node_id in set").