Do collapse leaves with other leaves.
In the previous commit (3ddfb39a), we did not collapse a node if
it was a leaf. But this is wrong. When there are many leaves
following each other, each very small, they should be collapsed
together.

This change allows leaves to collapse with other leaves, but if
a leaf ends up being the only one in a collapsed node, it
draws it normally instead of showing it collapsed.

This has a dramatic effect on trees like the ncbi. The amount
of graphical elements go from ~80k to around ~200, and the
speedup is enormous.

Also, the summary() function is quite fast the way it finds
names to represent a node. It is tricky to change it without
making the rendering of the tree much slower.