Draw a nodebox when we uncollapse a leaf.
Otherwise, even if we drew the leaf, the box that appeared on top
of it would say it was collapsed (when putting the cursor on it,
for example).

This makes it nicer to interact with in the gui. Still, we do not
send the node_id in this case, which can limit things like tags
and the available contextmenu operations. That seems to be a minor
problem.