Always download the newick of the modified tree.
The option to dowload the current tree, that was available from
the gui, would download the original newick if we tried to
download the full tree being in the main tree.

This had the advantage of getting the original, without any changes
made by the parsing or standardize(). And of being faster. The
disadvantage is that any modifications made to the tree would not
appear on the newick.

With this change, we go for the storing the changes version, which
is the less painful of the two options for me as a user.