cosmetics - refer to "node" as "branch" where appropriate in the context menu.
The way we refer to "nodes" can be confusing. I normally call a
"node" something with a content and children. But this is arguably
not standard terminology.

With this change we at least call some operations in the context
menu with names that are generally more clear, referring to branches
when the operations affect all descendants.