tree.pyx: Add class Children (list) that sets automatically the parent.
This way, whenever we do
  tree.children.append(node)
or
  tree.children += nodes
we automaticaly set the parent of all the nodes to tree.

So, no need to remember to set the parents, all while keeping the
same syntax as normal lists.

The performance tests show that it does not take more memory or
make it slower in any significant way.