Rerooting with better names and fix minor bug.
This change adds assertions and better names for the different
operations that we do when rerooting on a node.

The initial branch split did not copy the support value of the
original branch, and when we do multiple rerootings and the
tree had support values on those branches, we could have lost
them. This change fixes it.